super class WriteToFile made so files can know be written to a stranded folder in the users home directory, will be made if not present.
Logger class in FileHandling made so we can log actions taken in game, will be appended to a file called game.log, which is made if not present.

implemented logger to game class to log ever command written, second command word will print null if not present.

Fixed so the project can be run from any file not just the old SemesterProjekt class.